backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
PyMuPDF==1.24.10
pytesseract==0.3.13
Pillow==10.4.0
google-generativeai==0.8.3
sqlalchemy==2.0.35
alembic==1.13.3
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
